Add endpoint demonstration notebooks
This commit adds Jupyter Notebooks that demonstrate how the backend is interacted by simulating a client. Notably, the notebooks contain the note analyser functionality of the artefact, in which a user's subnote can be decrypted and analysed to produce a mood determination - which is based on a previously created wordlist by the user.